bot/pyproject.toml
[tool.poetry]
name = "mai"
version = "0.0.1"
description = ""
authors = ["Ghoul <you@example.com>"]
license = "GPL 3.0"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^5.4.1"
Pycord18n = "^1.0.0"
python-anilist = "^1.0.4"
animec = "^0.3.3"
aPRAW = "^0.6.9-alpha.0"
spotify = "^0.10.2"
osu-tools = "^0.2.13"
genshinstats = "^1.4.9"
humanize = "^3.11.0"
py-cord = {git = "https://github.com/Pycord-Development/pycord.git", rev = "master", extras = ["voice"]}
ossapi = "^2.2.0"
inquirerpy = "^0.3.0"

[tool.poetry.dev-dependencies]
aerich = "^0.5.8"
psutil = "^5.8.0"
youtube_dl = "^2021.6.6"
download = "^0.3.5"
watchgod = "^0.7"
aioredis = "^1.3.1"
update-checker = "^0.18.0"
Pycord18n = "^1.0.0"
black = "^21.9b0"
pre-commit = "^2.15.0"
mkdocs-material = "^7.3.3"
python-dateutil = "^2.8.2"
PycordUtils = "^0.0.2"
pyshorteners = "^1.0.1"
halo = "^0.0.31"
py-cpuinfo = "^8.0.0"
aiofiles = "^0.7.0"
rich = "^10.10.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.7"}
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
